+++ test/tx-data/buy.test.js
@@ -1,17 +1,19 @@
-import MinterSellAllCoinTxData from '../../src/tx-data/sell-all-coin';
+import MinterBuyTxData from '../../src/tx-data/buy';
 import {formatCoin} from '../../src/helpers';
 import decodeToArray from '../decode-to-array';
 
-describe('MinterBuyCoinTxData', () => {
+describe('MinterBuyTxData', () => {
     test('rlp encoded fields', () => {
-        const serializedTxData = (new MinterSellAllCoinTxData({
-            coin_to_sell: formatCoin('MNT'),
-            coin_to_buy: formatCoin('BELTCOIN'),
+        const serializedTxData = (new MinterBuyTxData({
+            coin_to_buy: formatCoin('MNT'),
+            value_to_buy: 10,
+            coin_to_sell: formatCoin('BELTCOIN'),
         })).serialize();
 
         expect(decodeToArray(serializedTxData))
             .toEqual([
                 [77, 78, 84, 0, 0, 0, 0, 0, 0, 0],
+                [10],
                 [66, 69, 76, 84, 67, 79, 73, 78, 0, 0],
             ]);
     });