+++ test/tx-data/sell.test.js
@@ -1,13 +1,13 @@
-import MinterBuyCoinTxData from '../../src/tx-data/buy-coin';
+import MinterSellTxData from '../../src/tx-data/sell';
 import {formatCoin} from '../../src/helpers';
 import decodeToArray from '../decode-to-array';
 
-describe('MinterBuyCoinTxData', () => {
+describe('MinterSellTxData', () => {
     test('rlp encoded fields', () => {
-        const serializedTxData = (new MinterBuyCoinTxData({
-            coin_to_buy: formatCoin('MNT'),
-            value_to_buy: 10,
-            coin_to_sell: formatCoin('BELTCOIN'),
+        const serializedTxData = (new MinterSellTxData({
+            coin_to_sell: formatCoin('MNT'),
+            value_to_sell: 10,
+            coin_to_buy: formatCoin('BELTCOIN'),
         })).serialize();
 
         expect(decodeToArray(serializedTxData))