+++ test/tx-data/set-candidate-on.test.js
@@ -2,7 +2,7 @@ import {toBuffer} from 'minterjs-util';
 import MinterSetCandidateOnTxData from '../../src/tx-data/set-candidate-on';
 import decodeToArray from '../decode-to-array';
 
-describe('MinterBuyCoinTxData', () => {
+describe('MinterSetCandidateOnTxData', () => {
     test('rlp encoded fields', () => {
         const serializedTxData = (new MinterSetCandidateOnTxData({
             pubkey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),