+++ test/tx-data/unbond.test.js
@@ -3,7 +3,7 @@ import MinterUnbondTxData from '../../src/tx-data/unbond';
 import {formatCoin} from '../../src/helpers';
 import decodeToArray from '../decode-to-array';
 
-describe('MinterBuyCoinTxData', () => {
+describe('MinterUnbondTxData', () => {
     test('rlp encoded fields', () => {
         const serializedTxData = (new MinterUnbondTxData({
             pubkey: toBuffer('Mpf9e036839a29f7fba2d5394bd489eda927ccb95acc99e506e688e4888082b3a3'),