+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.4.0 - 2018-08-14
+- **BREAKING** rename tx-data files: `sell-coin` -> `sell`, `sell-all-coin` -> `sell-all`, `buy-coin` -> `buy` 
+- add `converter.bipToPip` and `converter.pipToBip` alias methods
+
 ## 1.3.0 - 2018-08-11
 - **BREAKING** rename `unbound` to `unbond`, tx data and tx type are affected